Split @/api into @/auth @/notifier and @/common
