+++ packages/server/src/auth/api/index.ts
@@ -1,0 +1,2 @@
+export * from './auth'
+export * from './member'
