+++ packages/server/src/notifier/api/index.ts
@@ -1,10 +1,8 @@
 import { enumType } from 'nexus'
 import { NotificationKind } from 'nexus-prisma'
 
-export * from './auth'
 export * from './entitiySubscription'
 export * from './generalSubscription'
-export * from './member'
 export * from './notification'
 
 export const NotificationKindEnum = enumType(NotificationKind)
